% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutSeqRData.R
\docType{package}
\name{MutSeqRData}
\alias{MutSeqRData-package}
\alias{MutSeqRData}
\title{Supporting data for MutSeqR package}
\description{
\strong{Package:} MutSeqRData provides example data for MutSeqR.
 It includes the following objects:

\describe{
 \item{\strong{Example import_mut_data:}}{Data frame object containing list of sequenced sites for sample dna00996.1. Used as an example for importing tabular datasets with MutSeqR.}
\item{\strong{Example import_mut_data using Custom column names:}}{Data frame object containing list of sequenced sites for sample dna00996.1. Used as an example for importing tabular datasets with MutSeqR when dataset has custom column names.}
\item{\strong{Example import_vcf_data:}}{Block compressed VCF file containing list of sequenced sites for sample dna00996.1. Used as an example for importing VCF datasets with MutSeqR.}
\item{\strong{Example mutation data:}}{Data frame object containing list of sequenced sites for 24 samples. Unfiltered. Used as an example for filtering mutation data with MutSeqR.}
\item{\strong{Example mutation data filtered:}}{Data frame object containing the filtered list of sequenced sites for 24 samples. Used as an example for downstream MutSeqR functions.}
\item{\strong{Precalculated Depth at Base 6 Resolution:}}{Data frame object containing the pre-calculated total depth values for each sample at 6-base subtype resolution. Used as an example in calculate_mf()}
\item{\strong{Precalculated Depth at Base 12 Resolution:}}{Data frame object containing the pre-calculated total depth values for each sample at 12-base subtype resolution. Used as a reference for users.}
\item{\strong{Precalculated Depth at Base 96 Resolution:}}{Data frame object containing the pre-calculated total depth values for each sample at 96-base subtype resolution. Used as a reference for users.}
\item{\strong{Precalculated Depth at Base 192 Resolution:}}{Data frame object containing the pre-calculated total depth values for each sample at 192-base subtype resolution. Used as a reference for users.}}
}
\author{
\strong{Maintainer}: Matthew J. Meier \email{matthew.meier@hc-sc.gc.ca} (\href{https://orcid.org/0000-0001-8199-8754}{ORCID})

Authors:
\itemize{
  \item Annette E. Dodge \email{annette.dodge@hc-sc.gc.ca} (\href{https://orcid.org/0000-0002-0446-9055}{ORCID})
  \item Andrew Williams (\href{https://orcid.org/0000-0002-7637-7686}{ORCID})
  \item Danielle P.M. LeBlanc (\href{https://orcid.org/0000-0002-3847-8371}{ORCID})
  \item David M. Schuster (\href{https://orcid.org/0009-0001-6316-4358}{ORCID})
  \item Elena Esina (\href{https://orcid.org/0009-0002-3443-378X}{ORCID})
  \item Clint C. Valentine
  \item Jesse J. Salk
  \item Alexander Y. Maslov
  \item Christopher Bradley
  \item Carole L. Yauk (\href{https://orcid.org/0000-0002-6725-3454}{ORCID})
  \item Francesco Marchetti (\href{https://orcid.org/0000-0002-9435-4867}{ORCID})
}

Other contributors:
\itemize{
  \item Geronimo Matteo (\href{https://orcid.org/0000-0003-0819-4471}{ORCID}) [contributor]
}

}
\keyword{utilities}
